Naming conventions changed, launch options added, moved logger object
Changed name of char buffer for data entries in LogVehicleData()

Changed naming of VehicleLogger to CSV_Logger in all instances

Moved the index variable from main() to CommonMini, now a private variable inside of CSV_Logger

Moved the CSV_Log object and its instantiator into PlayerBase.

Added argument parcing option for the CSV log to enable it. By default it is disabled. It is now located in PlayerBase