Add data to OSI orientationRate.yaw
By using previous heading and the new heading
the orientation rate is calculated.
The orientation rate is set after the MoveAlongS function
which steps each object forward so we are sure
we are getting the new heading.
The heading rate is found in the position object along
with similar data, such as heading and velocity.