Fix controller bug in FollowTrajectoryAction #skip_bins
The ExternalController is supported to work also in additive mode, in
which case scenario actions are applied prior to callback for each
object. Then the object state can be manipulated within the scope of the
callback. The bug caused the action to always ignore any timing info of the
trajectory when a controller is attached, regardless of mode. Potenially
resulting in wrong longitudinal speed.