Fix macOS Big Sur issue #skip_bins
For some reason there is a limit on mac how close to screen upper border
the esmini window can be positioned. We noticed on Catalina that we had
to move it down from 30 to 50 pixels from top row (from Mojave). And now
on Big Sur the same problem arise. It seems that 60 pixels is OK.