Reserve Space key for esmini #skip_bins
In replayer Space key is used for paus/play. When camera is dismounted
from objects (key '4') a OSG view reset makes the camera move away.
Instead we want the possibility for the user to find a nice location for the
camera and run/step the scenario (using arrow and Space keys) from there.