Improve lane matching in XY2Road mapping #skip_bins
In mapping an arbitrary X,Y position to the road network, in terms of
road ID, lane ID, lane offset and s, it can be tricky to determine what
lane to choose. Especially in junctions with overlapping connecting
roads. This commit improves cases with one or multiple connected roads,
where the straightmost one is selected. The transition between roads should
now be smoother.