Support freespace distance measurement #skip_bins
Some conditions and actions offer a choice on how to measure distance:
1. freespace = false: Distance between the reference points of objects (low fidelity)
2. freespace = true: Distance between closest bounding box points (high fidelity)

So far esmini has accepted freespace=true without actually implementing it, leading to no effect but the same result as for freespace=false. Now it is implemented and applied accordingly.

Conditions with freespace option:
- TimeHeadwayCondition
- TimeToCollisionCondition

Actions with freespace option:
- DistanceCondition
- RelativeDistanceCondition
- LongitudinalDistanceAction
- (LateralDistanceAction - not implemented in esmini yet)