Make WASM example HTML more user friendly
This currently only works properly in Chrome based browsers and not in
Firefox since Firefox does not support the directory file system API
which is necessary for this use-case if we don't want to restrict the
scenario directory structure.
However, the embedded scenario also works with Firefox so it is not a
downgrade compared to the existing example.

Signed-off-by: Johannes Schmitz <johannes.schmitz1@gmail.com>